+++ internal/delivery/grpc/proto/antibruteforce.proto
@@ -1,0 +1,40 @@
+syntax = "proto3";
+
+package antibruteforce;
+option go_package = "./;pb";
+
+service AntiBruteForce {
+  rpc CheckAuthorization (AuthRequest) returns (AuthResponse) {}
+  rpc ResetBucket (ResetRequest) returns (ResetResponse) {}
+  rpc AddToBlacklist (ListRequest) returns (ListResponse) {}
+  rpc RemoveFromBlacklist (ListRequest) returns (ListResponse) {}
+  rpc AddToWhitelist (ListRequest) returns (ListResponse) {}
+  rpc RemoveFromWhitelist (ListRequest) returns (ListResponse) {}
+}
+
+message AuthRequest {
+  string login = 1;
+  string password = 2;
+  string ip = 3;
+}
+
+message AuthResponse {
+  bool ok = 1;
+}
+
+message ResetRequest {
+  string login = 1;
+  string ip = 2;
+}
+
+message ResetResponse {
+  bool success = 1;
+}
+
+message ListRequest {
+  string subnet = 1;
+}
+
+message ListResponse {
+  bool success = 1;
+}
