+++ internal/entity/user.go
@@ -1,0 +1,7 @@
+package entity
+
+type User struct {
+	ID       string `json:"id"`
+	Login    string `json:"login"`
+	Password string `json:"password"`
+}
