--- internal/logger/logger.go
@@ -1,21 +1,0 @@
-package logger
-
-import (
-	"github.com/sirupsen/logrus"
-)
-
-type Logger struct {
-	*logrus.Logger
-}
-
-func New(level string) *Logger {
-	logger := logrus.New()
-
-	lvl, err := logrus.ParseLevel(level)
-	if err != nil {
-		lvl = logrus.InfoLevel
-	}
-	logger.SetLevel(lvl)
-
-	return &Logger{logger}
-}
